test: Move RuntimeVerifier to K8sVerifier
This commit moves our RuntimeVerifier test to K8sVerifier, to be
executed on our 4.9, 4.19, and net-next test VMs.

The BPF programs are currently compiled to try and achieve the maximum
complexity possible on 4.9 (where e.g. BPF NodePort is not supported).
The Makefile will be extended in a subsequent patchset to include
max-complexity targets for 4.19 and net-next.

Signed-off-by: Paul Chaignon <paul@cilium.io>